Authorization complete

+++ backend-node/routes/posts.js
@@ -31,8 +31,9 @@ router.post("", checkAuth,multer({storage : storage}).single("image"),(req,res,n
     const post = new Post({
         title : req.body.title,
         content : req.body.content,
-        imagePath : url + "/images/" + req.file.filename
-    });
+        imagePath : url + "/images/" + req.file.filename,
+        creator : req.userData.userId
+        });
     post.save().then(CreatedPost => {
         res.status(201).json({
             message : 'Post added successfully!',
@@ -57,12 +58,16 @@ router.put("/:id",checkAuth, multer({storage : storage}).single("image"), (req, 
         _id: req.body.id,
         title : req.body.title,
         content : req.body.content,
-        imagePath: imagePath
+        imagePath: imagePath,
+        creator: req.userData.userId
     });
     console.log(post);
-    Post.updateOne({_id: req.params.id}, post).then(result => {
-       console.log(result);
-       res.status(200).json({message: "Updated Suceesfully"});  
+    Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post).then(result => {
+       if(result.nModified > 0){
+           res.status(200).json({message: "Updated Suceesfully"});  
+       }else{
+           res.status(401).json({ message:" You Does not create those post" })
+       }
     }).catch(error =>{
         console.log(error);
     })
@@ -107,10 +112,12 @@ router.get("/:id",(req,res,next)=>{
 
 router.delete("/:id",checkAuth,(req,res,next) => {
     console.log(req.params.id);
-    Post.deleteOne({_id : req.params.id}).then(result => {
-        console.log(result);
-        res.status(200).json({message : 'Post Deleted'});
-    })
+    Post.deleteOne({_id : req.params.id, creator: req.userData.userId}).then(result => {
+        if(result.n > 0){
+            res.status(200).json({message: "Deleted Suceesfully"});  
+        }else{
+            res.status(401).json({ message:" You Does not create those post" })
+        }    })
 });
 
 module.exports = router;
